+++ packages/core/src/index.ts
@@ -4,6 +4,7 @@ export * from "./abstract/Message.js"
 export * from "./abstract/Module.js"
 export * from "./abstract/ModulePaginator.js"
 export * from "./abstract/ReceivedInteraction.js"
+export * from "./abstract/Role.js"
 export * from "./abstract/Server.js"
 export * from "./abstract/User.js"
 
@@ -11,7 +12,7 @@ export * from "./abstract/User.js"
 export * from "./classes/Component.js"
 export * from "./classes/ComponentHandler.js"
 export * from "./classes/CrossBuild.js"
-export * from "./classes/OptionsHandler.js"
+export * from "./abstract/OptionsHandler.js"
 export * from "./classes/Paginator.js"
 
 // ==================== TYPINGS ====================
