+++ packages/guilded/src/index.ts
@@ -4,6 +4,7 @@ export * from "./GuildedModule"
 export * from "./classes/GuildedChannel"
 export * from "./classes/GuildedMessage"
 export * from "./classes/GuildedModulePaginator"
+export * from "./classes/GuildedOptionsHandler"
 export * from "./classes/GuildedReceivedMessage"
 export * from "./classes/GuildedServer"
 export * from "./classes/GuildedUser"
